A systems test using for all components
A systems test involving both stepper motors, servos and MIDI. Servo positions for damping and pick velocity are also included.
Uses both the AccellStepper and MIDI libraries that need to be down loaded separately (links in file).    

The stepper wheel system still needs to be fixed so it know what pick is going to be used.